Update to v1.c
Minor IPbus interface changes
Reconnection procedure improved to account delay for PMs to get ready.
Sides average time in TCM is showed for new (late November 2021) FW.
PMs and TCM counters FIFO buffers are read independently (HW count timer mode).

+++ PM.h
@@ -175,7 +175,7 @@ struct TypePM {
         static const quint8
             number = 24,
             addressDirect   =  0xC0;
-        quint16 FIFOload;
+        quint32 FIFOload;
         QDateTime newTime, oldTime = QDateTime::currentDateTime();
         union {
             quint32 New[number] = {0};
